修改短信功能
修改短信功能

+++ pcmss.dev/pcmss/src/main/java/com/quick/portal/sysUser/SysUserServiceImpl.java
@@ -22,6 +22,7 @@ package com.quick.portal.sysUser;
 import com.quick.core.base.SysBaseService;
 import com.quick.core.base.model.DataStore;
 import com.quick.portal.ldapmng.IUserLdapMngDao;
+import com.quick.portal.security.authority.metric.PropertiesUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -65,11 +66,13 @@ public class SysUserServiceImpl extends SysBaseService<SysUserDO> implements ISy
             sysUserDO.setUser_password("21232f297a57a5a743894a0e4a801fc3");
         sysUserDO.setRoles(String.join(",", sysUserDO.getRole_ids()));
         Integer val = sysUserDO.getUser_id();
-
-        if(val == null || val == 0) {
-            userLdapMngDao.saveUserLdapInfo(sysUserDO);
-        }else{
-            userLdapMngDao.updateUserLdapInfo(sysUserDO);
+        String ldapState = PropertiesUtil.getPropery("ldap.auth.set");
+        if("true".equals(ldapState)){
+            if(val == null || val == 0) {
+                userLdapMngDao.saveUserLdapInfo(sysUserDO);
+            }else{
+                userLdapMngDao.updateUserLdapInfo(sysUserDO);
+            }
         }
         return super.save(sysUserDO);
     }
@@ -90,7 +93,10 @@ public class SysUserServiceImpl extends SysBaseService<SysUserDO> implements ISy
         if(null != retList && !retList.isEmpty()){
             sysVO = retList.get(0);
         }
-        userLdapMngDao.removeUserLdapInfo(sysVO.getUser_name());
+        String ldapState = PropertiesUtil.getPropery("ldap.auth.set");
+        if("true".equals(ldapState)){
+            userLdapMngDao.removeUserLdapInfo(sysVO.getUser_name());
+        }
         dao.delete(sysUserDO);
         if (sysUserDO.getError_no() == 1)
             return ActionMsg.setOk("操作成功");
@@ -156,28 +162,30 @@ public class SysUserServiceImpl extends SysBaseService<SysUserDO> implements ISy
 
     public DataStore syncUserLdap(String userid){
         int uid  = Integer.valueOf(userid);
-        List<Map<String, Object>> retList = dao.getUserInfoByIds(uid);
-        int cnt = 0 ;
-        if(null !=retList && !retList.isEmpty()){
-            SysUserDO sysUserDO = null;
-            for(Map<String, Object> mp:retList){
-                sysUserDO = new SysUserDO();
-                sysUserDO.setUser_real_name(mp.get("user_real_name")==null?"":mp.get("user_real_name").toString());
-                sysUserDO.setUser_name(mp.get("user_name").toString());
-                sysUserDO.setUser_email(mp.get("user_email")==null?"":mp.get("user_email").toString());
-                sysUserDO.setUser_state(Integer.valueOf(mp.get("user_state").toString()));
-                sysUserDO.setUser_password(mp.get("user_password").toString());
-                sysUserDO.setUser_tel(mp.get("user_tel")==null?"":mp.get("user_tel").toString());
-                sysUserDO.setUser_id(Integer.valueOf(mp.get("user_id").toString()));
-                sysUserDO.setJob_name(mp.get("job_namne")==null?"":mp.get("job_namne").toString());
+        String ldapState = PropertiesUtil.getPropery("ldap.auth.set");
+        if("true".equals(ldapState)){
+            List<Map<String, Object>> retList = dao.getUserInfoByIds(uid);
+            int cnt = 0 ;
+            if(null !=retList && !retList.isEmpty()){
+                SysUserDO sysUserDO = null;
+                for(Map<String, Object> mp:retList){
+                    sysUserDO = new SysUserDO();
+                    sysUserDO.setUser_real_name(mp.get("user_real_name")==null?"":mp.get("user_real_name").toString());
+                    sysUserDO.setUser_name(mp.get("user_name").toString());
+                    sysUserDO.setUser_email(mp.get("user_email")==null?"":mp.get("user_email").toString());
+                    sysUserDO.setUser_state(Integer.valueOf(mp.get("user_state").toString()));
+                    sysUserDO.setUser_password(mp.get("user_password").toString());
+                    sysUserDO.setUser_tel(mp.get("user_tel")==null?"":mp.get("user_tel").toString());
+                    sysUserDO.setUser_id(Integer.valueOf(mp.get("user_id").toString()));
+                    sysUserDO.setJob_name(mp.get("job_namne")==null?"":mp.get("job_namne").toString());
 
-                cnt = userLdapMngDao.searchUserLdapCnt(sysUserDO);
-                if(cnt == 0){
-                    userLdapMngDao.saveUserLdapInfo(sysUserDO);
-                }else{
-                    userLdapMngDao.updateUserLdapInfo(sysUserDO);
+                    cnt = userLdapMngDao.searchUserLdapCnt(sysUserDO);
+                    if(cnt == 0){
+                        userLdapMngDao.saveUserLdapInfo(sysUserDO);
+                    }else{
+                        userLdapMngDao.updateUserLdapInfo(sysUserDO);
+                    }
                 }
-
             }
         }
         return ActionMsg.setOk("同步LDAP用户成功");
